requirements.txt
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==35.0.0
Flask==2.1.1
google-auth==2.6.3
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kubernetes==23.3.0
MarkupSafe==2.1.1
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
websocket-client==1.3.2
Werkzeug==2.1.1
zipp==3.8.0
uhashring==2.1
google-cloud-container==2.10.7
